Split request methods into endpoint classes
